TEST: Removing warnings from approx1 and approx2

+++ test/approx1.cpp
@@ -149,7 +149,7 @@ void approx1CubicTest(string pTestFile, const unsigned resultIdx, const af_inter
     bool ret = true;
 
     float max = real(outData[0]), min = real(outData[0]);
-    for(int i=1; i < nElems; ++i) {
+    for(int i=1; i < (int)nElems; ++i) {
         min = (real(outData[i]) < min) ? real(outData[i]) : min;
         max = (real(outData[i]) > max) ? real(outData[i]) : max;
     }
